updating references couchapp renamed map_reduce
