+++ .gitignore
@@ -6,7 +6,7 @@ compiled_handlebars.js
 #ignore generated folders
 release/
 dist
-couchapp_data/_attachments
+map_reduce_data/_attachments
 couchapp_minified/views
 couchapp_minified/lists
 couchapp_minified/_attachments
