+++ bower.json
@@ -39,13 +39,13 @@
     "bower.json",
     "bower_components",
     "build",
-    "couchapp_activities",
-    "couchapp_cleaning",
-    "couchapp_unused",
-    "couchapp_data",
-    "couchapp_export",
-    "couchapp_learnx",
-    "couchapp_lexicon",
+    "map_reduce_activities",
+    "map_reduce_cleaning",
+    "map_reduce_unused",
+    "map_reduce_data",
+    "map_reduce_export",
+    "map_reduce_learnx",
+    "map_reduce_lexicon",
     "couchapp_minified",
     "docs",
     "install_linux_download_and_set_up_fielddb_servers_for_new_developers_quick_start.sh",
