+++ docs/SoftwareDesignDescription.tex
@@ -174,7 +174,7 @@ https://github.com/FieldDB/FieldDB/tree/master/backbone\_client
 
 https://github.com/FieldDB/FieldDB/tree/master/angular\_client/modules/visualization
 
-https://github.com/FieldDB/FieldDB/tree/master/couchapp\_dev
+https://github.com/FieldDB/FieldDB/tree/master/map_reduce\_data
 
 \subsection{Search Client Component}
