+++ map_reduce_export/README.md
@@ -1,6 +1,6 @@
-This is the couchapp files needed to deploy the export widgets so they can be hosted on a couchdb, and embeded in other websites via an iframe.
+This is the map_reduce files needed to deploy the export widgets so they can be hosted on a couchdb, and embeded in other websites via an iframe.
 
-To build this couchapp run the script:
+To build this map_reduce run the script:
 
     $ cd FieldDB
     $ bash scripts/build_export_widgets.sh
