+++ map_reduce_lexicon/README.md
@@ -1,3 +1,3 @@
-This is a couchapp for the Lexicon viewer and generator
+This is a map_reduce for the Lexicon viewer and generator
 
-You can deploy this couchapp to a corpus to turn on the lexicon for that corpus. 
+You can deploy this map_reduce to a corpus to turn on the lexicon for that corpus.
