+++ map_reduce_wordcloud/README.md
@@ -1,3 +1,3 @@
-This is a couchapp for the Word Cloud viewer and generator
+This is a map_reduce for the Word Cloud viewer and generator
 
-You can deploy this couchapp to a corpus to turn on the word cloud features for that corpus. 
+You can deploy this map_reduce to a corpus to turn on the word cloud features for that corpus.
