+++ scripts/build_activity_feed.sh
@@ -16,10 +16,10 @@ backbone_client/node_modules/requirejs/bin/r.js -o  angular_client/modules/activ
 cp angular_client/modules/activity/release/js/main.js angular_client/modules/activity/client/libs/fielddb_activity_feed_widget.js
 cp angular_client/modules/activity/release/js/main.js backbone_client/activity/libs/fielddb_activity_feed_widget.js
 
-#putting the activity feed widgets into the minified couchapp
+#putting the activity feed widgets into the minified map_reduce
 cp angular_client/modules/activity/release/js/main.js couchapp_minified/_attachments/activity/libs/fielddb_activity_feed_widget.js
 
-# Putting activity feed files into the couchapp_activities which can host the activity feeds
-rm -rf couchapp_activities/_attachments
-cp -r angular_client/modules/activity/release couchapp_activities/_attachments
+# Putting activity feed files into the map_reduce_activities which can host the activity feeds
+rm -rf map_reduce_activities/_attachments
+cp -r angular_client/modules/activity/release map_reduce_activities/_attachments
