+++ scripts/build_template_databases_using_fielddb.sh
@@ -8,6 +8,6 @@ cd couchapp_minified
 erica push . $1"/new_corpus"
 
 echo "Deploy template activity feeds"
-cd ../couchapp_activities
+cd ../map_reduce_activities
 erica push . $1"/new_corpus_activity_feed"
 erica push . $1"/new_user_activity_feed"
